IGAPP-1329: Split button components web
