+++ web/src/components/FeedbackContainer.tsx
@@ -19,7 +19,7 @@ import { cmsApiBaseUrl } from '../constants/urls'
 import { RouteType } from '../routes'
 import { reportError } from '../utils/sentry'
 import Feedback, { Container } from './Feedback'
-import TextButton from './TextButton'
+import TextButton from './base/TextButton'
 
 const Text = styled.div`
   padding: 10px;
@@ -130,7 +130,7 @@ export const FeedbackContainer = (props: FeedbackContainerProps): ReactElement =
   return (
     <Container>
       <Text>{t('thanksMessage')}</Text>
-      {!!closeModal && !isSearchFeedback && <TextButton type='primary' onClick={closeModal} text={t('common:close')} />}
+      {!!closeModal && !isSearchFeedback && <TextButton onClick={closeModal} text={t('common:close')} />}
     </Container>
   )
 }
