+++ web/src/components/FeedbackSearch.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components'
 import { FeedbackType } from 'api-client'
 
 import FeedbackContainer from './FeedbackContainer'
-import TextButton from './TextButton'
+import TextButton from './base/TextButton'
 
 const Container = styled.div`
   display: flex;
@@ -43,7 +43,7 @@ const FeedbackSearch = ({ cityCode, languageCode, query, resultsFound }: Feedbac
   }
   return (
     <Container>
-      <TextButton type='primary' onClick={handleFeedbackLinkClicked} text={t('informationNotFound')} />
+      <TextButton onClick={handleFeedbackLinkClicked} text={t('informationNotFound')} />
     </Container>
   )
 }
